+++ feature-recipe-list/src/main/java/com/github/belyakovleonid/feature_recipe_list/presentation/RecipesListViewModel.kt
@@ -23,8 +23,8 @@ class RecipeListViewModel @Inject constructor(
 
     private fun loadMoreRecipes() {
         viewModelScope.launch {
-//            val result = recipesInteractor.loadMoreRecipes()
-//            mutableState.value = transmitResultToState(result)
+            val result = recipesInteractor.loadMoreRecipes()
+            mutableState.value = transmitResultToState(result)
         }
     }
