+++ feature-weight-track/src/main/java/com/github/belyakovleonid/feature_weight_track/base/data/local/WeightGoalDao.kt
@@ -1,10 +1,10 @@
-package com.github.belyakovleonid.feature_weight_track.root.data.local
+package com.github.belyakovleonid.feature_weight_track.base.data.local
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.github.belyakovleonid.feature_weight_track.root.data.local.model.WeightGoalEntity
+import com.github.belyakovleonid.feature_weight_track.base.data.local.model.WeightGoalEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
@@ -20,7 +20,6 @@ interface WeightGoalDao {
     suspend fun getWeightGoal(): WeightGoalEntity?
 
     companion object {
-
         private const val SELECT_GOAL_QUERY = """
             SELECT * FROM ${WeightGoalEntity.WEIGHT_GOAL_TABLE_NAME}
             WHERE id = ${WeightGoalEntity.THE_ONLY_ID}
