+++ feature-weight-track/src/main/java/com/github/belyakovleonid/feature_weight_track/base/data/local/model/WeightGoalEntity.kt
@@ -1,9 +1,9 @@
-package com.github.belyakovleonid.feature_weight_track.root.data.local.model
+package com.github.belyakovleonid.feature_weight_track.base.data.local.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.github.belyakovleonid.feature_weight_track.root.data.local.model.WeightGoalEntity.Companion.WEIGHT_GOAL_TABLE_NAME
-import com.github.belyakovleonid.feature_weight_track.root.domain.model.WeightGoal
+import com.github.belyakovleonid.feature_weight_track.base.data.local.model.WeightGoalEntity.Companion.WEIGHT_GOAL_TABLE_NAME
+import com.github.belyakovleonid.feature_weight_track.base.domain.model.WeightGoal
 
 @Entity(tableName = WEIGHT_GOAL_TABLE_NAME)
 data class WeightGoalEntity(
