+++ feature-weight-track/src/main/java/com/github/belyakovleonid/feature_weight_track/base/di/WeightTrackModule.kt
@@ -4,9 +4,11 @@ import android.content.Context
 import androidx.room.Room
 import com.github.belyakovleonid.core.di.FragmentScope
 import com.github.belyakovleonid.core.viewmodel.AssistedVMFactory
-import com.github.belyakovleonid.feature_weight_track.root.data.WeightTrackRepositoryImpl
-import com.github.belyakovleonid.feature_weight_track.root.data.local.WeightTrackDatabase
-import com.github.belyakovleonid.feature_weight_track.root.domain.WeightTrackRepository
+import com.github.belyakovleonid.feature_weight_track.base.data.WeightGoalRepositoryImpl
+import com.github.belyakovleonid.feature_weight_track.base.data.WeightTrackRepositoryImpl
+import com.github.belyakovleonid.feature_weight_track.base.data.local.WeightDatabase
+import com.github.belyakovleonid.feature_weight_track.base.domain.WeightGoalRepository
+import com.github.belyakovleonid.feature_weight_track.base.domain.WeightTrackRepository
 import com.github.belyakovleonid.feature_weight_track.weightpicker.presentation.WeightPickerViewModel
 import com.github.belyakovleonid.feature_weight_track.weightpicker.presentation.params.WeightPickerParams
 import dagger.Binds
@@ -21,26 +23,29 @@ object WeightTrackModule {
     @FragmentScope
     fun provideWeightTrackDatabase(
         context: Context
-    ): WeightTrackDatabase = Room.databaseBuilder(
+    ): WeightDatabase = Room.databaseBuilder(
         context,
-        WeightTrackDatabase::class.java,
+        WeightDatabase::class.java,
         "weight_track_database"
     ).build()
 
     @JvmStatic
     @Provides
-    fun provideWeightTrackDao(database: WeightTrackDatabase) = database.getWeightTrackDao()
+    fun provideWeightTrackDao(database: WeightDatabase) = database.getWeightTrackDao()
 
     @JvmStatic
     @Provides
-    fun provideWeightGoalDao(database: WeightTrackDatabase) = database.getWeightGoalDao()
+    fun provideWeightGoalDao(database: WeightDatabase) = database.getWeightGoalDao()
 }
 
 @Module
 interface WeightTrackBindsModule {
 
     @Binds
-    fun bindsWeightRepository(repository: WeightTrackRepositoryImpl): WeightTrackRepository
+    fun bindsTrackRepository(impl: WeightTrackRepositoryImpl): WeightTrackRepository
+
+    @Binds
+    fun bindsGoalRepository(impl: WeightGoalRepositoryImpl): WeightGoalRepository
 
     @Binds
     @FragmentScope
