+++ feature-weight-track/src/main/java/com/github/belyakovleonid/feature_weight_track/base/domain/model/WeightGoal.kt
@@ -1,0 +1,5 @@
+package com.github.belyakovleonid.feature_weight_track.base.domain.model
+
+data class WeightGoal(
+    val weight: Float
+)
