+++ feature-weight-track/src/main/java/com/github/belyakovleonid/feature_weight_track/base/domain/model/WeightTrack.kt
@@ -1,4 +1,4 @@
-package com.github.belyakovleonid.feature_weight_track.root.domain.model
+package com.github.belyakovleonid.feature_weight_track.base.domain.model
 
 import java.time.LocalDate
