--- feature-weight-track/src/main/java/com/github/belyakovleonid/feature_weight_track/root/domain/model/WeightGoal.kt
@@ -1,5 +1,0 @@
-package com.github.belyakovleonid.feature_weight_track.root.domain.model
-
-data class WeightGoal(
-    val weight: Float
-)
